+++ R/DataLoadingSaving.R
@@ -44,7 +44,7 @@
 #'                                     instance. Requires read permissions to this database. On SQL
 #'                                     Server, this should specify both the database and the schema,
 #'                                     so for example 'cdm_instance.dbo'.
-#' @param oracleTempSchema    DEPRECATED: use \code{tempEmulationSchema} instead.
+#' @param oracleTempSchema    DEPRECATED: use `tempEmulationSchema` instead.
 #' @param tempEmulationSchema Some database platforms like Oracle and Impala do not truly support temp tables. To
 #'                            emulate temp tables, provide a schema with write privileges where temp tables
 #'                            can be created.