+++ tests/testthat/test-psFunctions.R
@@ -199,3 +199,12 @@ test_that("Error messages for wrong input", {
 })
 
 
+test_that("IPTW ATT", {
+  rowId <- 1:5
+  treatment <- c(1, 0, 1, 0, 1)
+  propensityScore <- c(0.1, 0.2, 0.3, 0.4, 0.5)
+  data <- data.frame(rowId = rowId, treatment = treatment, propensityScore = propensityScore)
+  w <- computeWeights(data, estimator = "att")
+  wGoldStandard <- mean(treatment == 1)*treatment + mean(treatment == 0) * (1 - treatment) * propensityScore / (1 - propensityScore)
+  expect_equal(w, wGoldStandard)
+})